prob 101 initial commit
Both move functions appear to be working correctly base on the problem
definition. The pile functions have been started but they were abandoned
when I realized I missed a use case for the move onto function,
specifically that the directions don't say to move block B back to its
initial position if it's already on a stack. So by my understanding the
move onto function can stack as long as the block we're moving onto is
at the top of an existing stack.