prob 101 morning work
Re-read the problem description and realized I am worry about a
condition case that is not allowed. Specifically I have been working
with the issue of both blocks in the same stack and if they flip order
then things go haywire. This is by definition an illegal move so I don't
need to worry about it, but I do have to do a check to see if the two
blocks being moved are in the same stack. Moves where the blocks share a
stack are considered illegal so I added another variable to the block
struct so each block will know which stack it's in then a simple `if`
statement will resolve that illegal move.